backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.7
openai==1.6.1
requests==2.31.0
python-dotenv==1.0.0